Improve genzer.vscode role to only try installing new extensions
Previously, the task calls `code --install-extension` to install every
extension in the list, if the extension is already installed, the task
simply is considered unchanged.

However, this is slow as hell. So this commit improves that by only
attemp to install the extensions which is not installed. The task simply
obtain the list of installed extensions and then filter out the ones are
not in the list.